refactor: remove orphan files and slim FileExplorer to live helpers
