+++ src/components/common/index.ts
@@ -1,4 +1,3 @@
-export * from './SearchInput';
 export * from './linkBehavior';
 export * from './WatchlistButton';
 export * from './DataTable';
