+++ src/components/miners/index.ts
@@ -1,5 +1,4 @@
 export { default as MinerActivity } from './MinerActivity';
-export { default as MinerComparisonRadar } from './MinerComparisonRadar';
 export { default as MinerInsightsCard } from './MinerInsightsCard';
 export { default as MinerOpenDiscoveryIssuesByRepo } from './MinerOpenDiscoveryIssuesByRepo';
 export { default as MinerPRsTable } from './MinerPRsTable';
